refactor automation
