+++ .github/workflows/update_prices.yaml
@@ -17,8 +17,8 @@ jobs:
       - name: Build publisher images locally
         run: |
           docker build . --target test -t 42labs/empiric-publisher:test
-          docker build publisher/sample-publisher/all/ -t publish-all --build-arg EMPIRIC_PUBLISHER_BASE_IMAGE_TAG=test
-          docker build publisher/sample-publisher/coinbase/ -t publish-coinbase --build-arg EMPIRIC_PUBLISHER_BASE_IMAGE_TAG=test
+          docker build stagecoach/jobs/publishers/publish-all/ -t publish-all --build-arg EMPIRIC_PUBLISHER_BASE_IMAGE_TAG=test
+          docker build stagecoach/jobs/publishers/coinbase/ -t publish-coinbase --build-arg EMPIRIC_PUBLISHER_BASE_IMAGE_TAG=test
 
       - name: Post updated prices
         env:
@@ -30,7 +30,8 @@ jobs:
           FTX_API_SECRET: ${{ secrets.FTX_API_SECRET }}
         run: |
           docker run \
-            --env-file publisher/sample-publisher/all/.env\
+            --env-file stagecoach/jobs/publishers/publish-all/.env\
+            -p 9000:8080
             -e __EMPIRIC_STAGING_ENV__=TRUE\
             -e __EMPIRIC_PUBLISHER_EXIT_ON_ERROR__=TRUE\
             -e PUBLISHER_ADDRESS=2997231508090948287810421563528272582753036618878708141844804758659373413510\
@@ -41,13 +42,15 @@ jobs:
             -e FTX_API_KEY=${FTX_API_KEY}\
             -e FTX_API_SECRET=${FTX_API_SECRET}\
             publish-all
+          curl -XPOST "http://localhost:9000/2015-03-31/functions/function/invocations" -d '{}'
           docker run \
-            --env-file publisher/sample-publisher/coinbase/.env\
+            --env-file stagecoach/jobs/publishers/coinbase/.env\
             -e __EMPIRIC_STAGING_ENV__=TRUE\
             -e PUBLISHER_ADDRESS=2997231508090948287810421563528272582753036618878708141844804758659373413510\
             -e PUBLISHER_PRIVATE_KEY=${PUBLISHER_PRIVATE_KEY}\
             -e COINBASE_API_SECRET=${COINBASE_API_SECRET}\
             -e COINBASE_API_KEY=${COINBASE_API_KEY}\
             -e COINBASE_API_PASSPHRASE=${COINBASE_API_PASSPHRASE}\
+            -p 9001:8080
             publish-coinbase
-
+          curl -XPOST "http://localhost:9001/2015-03-31/functions/function/invocations" -d '{}'
