+++ empiric-package/empiric/core/__init__.py
@@ -1,3 +1,4 @@
 from empiric.core.client import EmpiricClient
 from empiric.core.contract import Contract
 from empiric.core.entry import Entry
+from empiric.core.types import AggregationMode
