+++ empiric-package/empiric/publisher/fetchers/thegraph.py
@@ -7,7 +7,7 @@ import requests
 from aiohttp import ClientSession
 from empiric.core.entry import Entry
 from empiric.publisher.assets import EmpiricAsset, EmpiricSpotAsset
-from empiric.publisher.base import PublisherInterfaceT
+from empiric.publisher.types import PublisherInterfaceT
 
 logger = logging.getLogger(__name__)
