--- publisher/manage-deployment/initialize_lightsail.sh
@@ -1,14 +1,0 @@
-sudo apt-get update
-sudo apt-get install -y \
-    ca-certificates \
-    curl \
-    gnupg \
-    lsb-release \
-    apt-transport-https \
-    software-properties-common
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu focal stable"
-apt-cache policy docker-ce
-sudo apt install -y docker-ce
-sudo usermod -aG docker ${USER}
-(crontab -l 2>/dev/null; echo "2-59/5 * * * * docker build all/ -t all && docker run --rm --env-file all/.env --env-file all/.secrets.env all > all/log.txt 2>&1") | crontab -
