--- publisher/monitor/balance/Dockerfile
@@ -1,5 +1,0 @@
-ARG EMPIRIC_PUBLISHER_BASE_IMAGE_TAG=0.8.1
-FROM 42labs/empiric-publisher:$EMPIRIC_PUBLISHER_BASE_IMAGE_TAG
-
-COPY monitor-balance.py ./monitor-balance.py
-CMD python monitor-balance.py
