--- publisher/monitor/price/.env
@@ -1,2 +1,0 @@
-BETTERUPTIME_ID=RVXYz4tMHb7wnr9CTKFTuRES
-SLACK_CHANNEL_ID=C03L872TG2F
