--- publisher/sample-publisher/all/.env
@@ -1,3 +1,0 @@
-PUBLISHER=empiric
-PUBLISHER_ADDRESS=2885948362156884910017225078393335267371301698754605063165748466266191873667
-BETTERUPTIME_ID=eLy7zigidGbx5s6jnsfQiqJQ
