--- publisher/sample-publisher/integration/.env
@@ -1,2 +1,0 @@
-PUBLISHER=empiric
-PUBLISHER_ADDRESS=3053905924911392281185613487163536321775910383295124523151415438844445678030
