+++ stagecoach/jobs/publishers/coinbase/coinbase_publisher.py
