+++ stagecoach/jobs/publishers/examples/.env
@@ -1,0 +1,2 @@
+PUBLISHER=empiric
+PUBLISHER_ADDRESS=1642826981151527159562401489136351744791574826801407816903513204997139632893
