+++ stagecoach/jobs/publishers/examples/integration_publisher.py
