+++ stagecoach/jobs/publishers/examples/publish_all.py
