+++ stagecoach/jobs/publishers/examples/register_publishers.py
