+++ stagecoach/monitors/balance/Dockerfile
